chore: add back two utils used in i18n and move to relevant dir
